+++ docs/api/tianshou.env.rst
@@ -5,3 +5,8 @@ tianshou.env
    :members:
    :undoc-members:
    :show-inheritance:
+
+.. automodule:: tianshou.env.worker
+   :members:
+   :undoc-members:
+   :show-inheritance:
