+++ docs/tutorials/tictactoe.rst
@@ -158,7 +158,6 @@ Tianshou already provides some builtin classes for multi-agent learning. You can
     ===x _ o x _ _===
     ===x _ _ _ x x===
     =================
-    >>> collector.close()
 
 Random agents perform badly. In the above game, although agent 2 wins finally, it is clear that a smart agent 1 would place an ``x`` at row 4 col 4 to win directly. 
 
@@ -175,7 +174,7 @@ So let's start to train our Tic-Tac-Toe agent! First, import some required modul
     from copy import deepcopy
     from torch.utils.tensorboard import SummaryWriter
 
-    from tianshou.env import VectorEnv
+    from tianshou.env import DummyVectorEnv
     from tianshou.utils.net.common import Net
     from tianshou.trainer import offpolicy_trainer
     from tianshou.data import Collector, ReplayBuffer
@@ -220,8 +219,7 @@ The explanation of each Tianshou class/function will be deferred to their first 
                             help='the path of opponent agent pth file for resuming from a pre-trained agent')
         parser.add_argument('--device', type=str,
                             default='cuda' if torch.cuda.is_available() else 'cpu')
-        args = parser.parse_known_args()[0]
-        return args
+        return parser.parse_args()
 
 .. sidebar:: The relationship between MultiAgentPolicyManager (Manager) and BasePolicy (Agent)
 
@@ -290,15 +288,14 @@ With the above preparation, we are close to the first learned agent. The followi
         collector = Collector(policy, env)
         result = collector.collect(n_episode=1, render=args.render)
         print(f'Final reward: {result["rew"]}, length: {result["len"]}')
-        collector.close()
     if args.watch:
         watch(args)
         exit(0)
 
     # ======== environment setup =========
     env_func = lambda: TicTacToeEnv(args.board_size, args.win_size)
-    train_envs = VectorEnv([env_func for _ in range(args.training_num)])
-    test_envs = VectorEnv([env_func for _ in range(args.test_num)])
+    train_envs = DummyVectorEnv([env_func for _ in range(args.training_num)])
+    test_envs = DummyVectorEnv([env_func for _ in range(args.test_num)])
     # seed
     np.random.seed(args.seed)
     torch.manual_seed(args.seed)
@@ -350,9 +347,6 @@ With the above preparation, we are close to the first learned agent. The followi
         args.batch_size, train_fn=train_fn, test_fn=test_fn,
         stop_fn=stop_fn, save_fn=save_fn, writer=writer,
         test_in_train=False)
-
-    train_collector.close()
-    test_collector.close()
 
     agent = policy.policies[args.agent_id - 1]
     # let's watch the match!
