+++ examples/box2d/bipedal_hardcore_sac.py
@@ -39,8 +39,7 @@ def get_args():
     parser.add_argument(
         '--device', type=str,
         default='cuda' if torch.cuda.is_available() else 'cpu')
-    args = parser.parse_known_args()[0]
-    return args
+    return parser.parse_args()
 
 
 class EnvWrapper(object):
@@ -136,7 +135,6 @@ def test_sac_bipedal(args=get_args()):
         args.step_per_epoch, args.collect_per_step, args.test_num,
         args.batch_size, stop_fn=IsStop, save_fn=save_fn, writer=writer)
 
-    test_collector.close()
     if __name__ == '__main__':
         pprint.pprint(result)
         # Let's watch its performance!
@@ -144,7 +142,6 @@ def test_sac_bipedal(args=get_args()):
         collector = Collector(policy, env)
         result = collector.collect(n_episode=16, render=args.render)
         print(f'Final reward: {result["rew"]}, length: {result["len"]}')
-        collector.close()
 
 
 if __name__ == '__main__':
