+++ examples/mujoco/mujoco/assets/point.xml
