+++ test/base/test_collector.py
@@ -2,7 +2,7 @@ import numpy as np
 from torch.utils.tensorboard import SummaryWriter
 
 from tianshou.policy import BasePolicy
-from tianshou.env import VectorEnv, SubprocVectorEnv, AsyncVectorEnv
+from tianshou.env import DummyVectorEnv, SubprocVectorEnv
 from tianshou.data import Collector, Batch, ReplayBuffer
 
 if __name__ == '__main__':
@@ -12,14 +12,24 @@ else:  # pytest
 
 
 class MyPolicy(BasePolicy):
-    def __init__(self, dict_state=False):
+    def __init__(self, dict_state: bool = False, need_state: bool = True):
+        """
+        :param bool dict_state: if the observation of the environment is a dict
+        :param bool need_state: if the policy needs the hidden state (for RNN)
+        """
         super().__init__()
         self.dict_state = dict_state
+        self.need_state = need_state
 
     def forward(self, batch, state=None):
+        if self.need_state:
+            if state is None:
+                state = np.zeros((len(batch.obs), 2))
+            else:
+                state += 1
         if self.dict_state:
-            return Batch(act=np.ones(len(batch.obs['index'])))
-        return Batch(act=np.ones(len(batch.obs)))
+            return Batch(act=np.ones(len(batch.obs['index'])), state=state)
+        return Batch(act=np.ones(len(batch.obs)), state=state)
 
     def learn(self):
         pass
@@ -66,32 +76,32 @@ def test_collector():
     env_fns = [lambda x=i: MyTestEnv(size=x, sleep=0) for i in [2, 3, 4, 5]]
 
     venv = SubprocVectorEnv(env_fns)
-    dum = VectorEnv(env_fns)
+    dum = DummyVectorEnv(env_fns)
     policy = MyPolicy()
     env = env_fns[0]()
     c0 = Collector(policy, env, ReplayBuffer(size=100, ignore_obs_next=False),
                    logger.preprocess_fn)
     c0.collect(n_step=3)
-    assert np.allclose(c0.buffer.obs[:4], np.expand_dims(
-                                                    [0, 1, 0, 1], axis=-1))
-    assert np.allclose(c0.buffer[:4].obs_next, np.expand_dims(
-                                                [1, 2, 1, 2], axis=-1))
+    assert np.allclose(c0.buffer.obs[:4],
+                       np.expand_dims([0, 1, 0, 1], axis=-1))
+    assert np.allclose(c0.buffer[:4].obs_next,
+                       np.expand_dims([1, 2, 1, 2], axis=-1))
     c0.collect(n_episode=3)
-    assert np.allclose(c0.buffer.obs[:10], np.expand_dims(
-                                [0, 1, 0, 1, 0, 1, 0, 1, 0, 1], axis=-1))
-    assert np.allclose(c0.buffer[:10].obs_next, np.expand_dims(
-                                    [1, 2, 1, 2, 1, 2, 1, 2, 1, 2], axis=-1))
+    assert np.allclose(c0.buffer.obs[:10],
+                       np.expand_dims([0, 1, 0, 1, 0, 1, 0, 1, 0, 1], axis=-1))
+    assert np.allclose(c0.buffer[:10].obs_next,
+                       np.expand_dims([1, 2, 1, 2, 1, 2, 1, 2, 1, 2], axis=-1))
     c0.collect(n_step=3, random=True)
     c1 = Collector(policy, venv, ReplayBuffer(size=100, ignore_obs_next=False),
                    logger.preprocess_fn)
     c1.collect(n_step=6)
     assert np.allclose(c1.buffer.obs[:11], np.expand_dims(
-                            [0, 1, 0, 1, 2, 0, 1, 0, 1, 2, 3], axis=-1))
-    assert np.allclose(c1.buffer[:11].obs_next, np.expand_dims([
-                                1, 2, 1, 2, 3, 1, 2, 1, 2, 3, 4], axis=-1))
+        [0, 1, 0, 1, 2, 0, 1, 0, 1, 2, 3], axis=-1))
+    assert np.allclose(c1.buffer[:11].obs_next, np.expand_dims(
+        [1, 2, 1, 2, 3, 1, 2, 1, 2, 3, 4], axis=-1))
     c1.collect(n_episode=2)
-    assert np.allclose(c1.buffer.obs[11:21], np.expand_dims(
-                            [0, 1, 2, 3, 4, 0, 1, 0, 1, 2], axis=-1))
+    assert np.allclose(c1.buffer.obs[11:21],
+                       np.expand_dims([0, 1, 2, 3, 4, 0, 1, 0, 1, 2], axis=-1))
     assert np.allclose(c1.buffer[11:21].obs_next,
                        np.expand_dims([1, 2, 3, 4, 5, 1, 2, 1, 2, 3], axis=-1))
     c1.collect(n_episode=3, random=True)
@@ -116,7 +126,7 @@ def test_collector_with_async():
     env_fns = [lambda x=i: MyTestEnv(size=x, sleep=0.1, random_sleep=True)
                for i in env_lens]
 
-    venv = AsyncVectorEnv(env_fns)
+    venv = SubprocVectorEnv(env_fns, wait_num=len(env_fns) - 1)
     policy = MyPolicy()
     c1 = Collector(policy, venv,
                    ReplayBuffer(size=1000, ignore_obs_next=False),
@@ -129,8 +139,6 @@ def test_collector_with_async():
     size = len(c1.buffer)
     obs = c1.buffer.obs[:size]
     done = c1.buffer.done[:size]
-    print(env_id[:size])
-    print(obs)
     obs_ground_truth = []
     i = 0
     while i < size:
@@ -165,7 +173,7 @@ def test_collector_with_dict_state():
     c0.collect(n_episode=2)
     env_fns = [lambda x=i: MyTestEnv(size=x, sleep=0, dict_state=True)
                for i in [2, 3, 4, 5]]
-    envs = VectorEnv(env_fns)
+    envs = DummyVectorEnv(env_fns)
     envs.seed(666)
     obs = envs.reset()
     assert not np.isclose(obs[0]['rand'], obs[1]['rand'])
@@ -185,7 +193,6 @@ def test_collector_with_dict_state():
                    Logger.single_preprocess_fn)
     c2.collect(n_episode=[0, 0, 0, 10])
     batch, _ = c2.buffer.sample(10)
-    print(batch['obs_next']['index'])
 
 
 def test_collector_with_ma():
@@ -202,7 +209,7 @@ def test_collector_with_ma():
     assert np.asanyarray(r).size == 1 and r == 4.
     env_fns = [lambda x=i: MyTestEnv(size=x, sleep=0, ma_rew=4)
                for i in [2, 3, 4, 5]]
-    envs = VectorEnv(env_fns)
+    envs = DummyVectorEnv(env_fns)
     c1 = Collector(policy, envs, ReplayBuffer(size=100),
                    Logger.single_preprocess_fn, reward_metric=reward_metric)
     r = c1.collect(n_step=10)['rew']
@@ -227,7 +234,6 @@ def test_collector_with_ma():
     r = c2.collect(n_episode=[0, 0, 0, 10])['rew']
     assert np.asanyarray(r).size == 1 and r == 4.
     batch, _ = c2.buffer.sample(10)
-    print(batch['obs_next'])
 
 
 if __name__ == '__main__':
