+++ tianshou/__init__.py
@@ -1,7 +1,7 @@
 from tianshou import data, env, utils, policy, trainer, \
     exploration
 
-__version__ = '0.2.5'
+__version__ = '0.2.6'
 __all__ = [
     'env',
     'data',
